requirements.txt
flask==3.0.3
line-bot-sdk==3.11.0
yfinance==0.2.40
requests==2.32.3
pytz==2024.1
gunicorn==22.0.0
